+++ lib/td_dd_web/controllers/search_controller.ex
@@ -57,36 +57,41 @@ defmodule TdDdWeb.SearchController do
   end
 
   def search_grants(conn, params) do
-    search(conn, params, :by_permissions)
+    claims = conn.assigns[:current_resource]
+    manage_permission = can?(claims, manage(Grant))
+    user_permissions = %{manage_grants: manage_permission}
+
+    %{total: total} = response = search(conn, params, :by_permissions)
+
+    conn
+    |> put_resp_header("x-total-count", "#{total}")
+    |> render("search.json", search_assigns(response, user_permissions))
   end
 
   def search_my_grants(conn, params) do
-    search(conn, params, :by_user)
+    claims = conn.assigns[:current_resource]
+    manage_permission = can?(claims, manage(Grant))
+    user_permissions = %{manage_grants: manage_permission}
+
+    %{total: total} = response = search(conn, params, :by_user)
+
+    conn
+    |> put_resp_header("x-total-count", "#{total}")
+    |> render("search.json", search_assigns(response, user_permissions))
+  end
+
+  defp search(_conn, %{"scroll_id" => _scroll_id, "scroll" => _scroll} = params, _by) do
+    Search.scroll_grants(params)
   end
 
   defp search(conn, params, by) do
     page = Map.get(params, "page", 0)
     size = Map.get(params, "size", 20)
     claims = conn.assigns[:current_resource]
-    manage_permission = can?(claims, manage(%{"resource_type" => "grant"}))
-    user_permissions = %{manage_grants: manage_permission}
 
-    %{
-      results: grants,
-      aggregations: aggregations,
-      total: total
-    } =
-      params
-      |> Map.drop(["page", "size"])
-      |> search(claims, page, size, by)
-
-    conn
-    |> put_resp_header("x-total-count", "#{total}")
-    |> render("search.json",
-      grants: grants,
-      filters: aggregations,
-      user_permissions: [user_permissions]
-    )
+    params
+    |> Map.drop(["page", "size"])
+    |> search(claims, page, size, by)
   end
 
   defp search(params, claims, page, size, :by_user) do
@@ -95,5 +100,13 @@ defmodule TdDdWeb.SearchController do
 
   defp search(params, claims, page, size, :by_permissions) do
     Search.search(params, claims, page, size)
+  end
+
+  defp search_assigns(%{results: grants, scroll_id: scroll_id}, _user_permissions) do
+    [grants: grants, scroll_id: scroll_id]
+  end
+
+  defp search_assigns(%{results: grants, aggregations: aggregations}, user_permissions) do
+    [grants: grants, filters: aggregations, user_permissions: user_permissions]
   end
 end